+++ gruntfile.js
@@ -4,40 +4,30 @@ module.exports = function (grunt) {
         clean: ['build'],
         babel: {
             options: {
-                sourceMap: false,
+                sourceMap: false
             },
             dist: {
                 files: [{
                     expand: true,
                     cwd: './lib',
-                    src: ['*.js'],
+                    src: ['**/*.js'],
                     dest: 'build',
                     ext: '.js'
                 }]
             }
         },
-        mocha_istanbul: {
-            coverage: {
-                src: ['test/*.spec.js'],
-                options: {
-                    scriptPath: require.resolve('isparta/bin/isparta'),
-                    reporter: 'spec',
-                    mochaOptions: ['--compilers', 'js:babel/register', '--recursive'],
-                    require: ['should']
-                }
-            }
-        },
         watch: {
             dist: {
-                files: ['./lib/*.js'],
+                files: ['./lib/**/*.js'],
                 tasks: ['babel:dist']
             }
         },
         eslint: {
-            options: {
-                parser: 'babel-eslint'
-            },
-            target: ['index.js']
+            target: [
+                'index.js',
+                'launcher.js',
+                'lib/**/*.js'
+            ]
         },
         contributors: {
             options: {
@@ -54,12 +44,11 @@ module.exports = function (grunt) {
 
     require('load-grunt-tasks')(grunt);
     grunt.registerTask('default', ['build']);
-    grunt.registerTask('build', 'Build wdio-sauce-service', function () {
+    grunt.registerTask('build', 'Build wdio-appium-service', function () {
         grunt.task.run([
             'eslint',
             'clean',
-            'babel',
-            'mocha_istanbul'
+            'babel'
         ]);
     });
     grunt.registerTask('release', 'Bump and tag version', function (type) {
