+++ index.js
@@ -1,0 +1,4 @@
+/**
+ * no service required
+ */
+module.exports = {};
