+++ package.json
@@ -33,19 +33,18 @@
     "babel-preset-es2015": "^6.6.0",
     "babel-register": "^6.7.2",
     "eslint": "^3.3.1",
-    "eslint-plugin-node": "^2.0.0",
+    "eslint-config-standard": "^5.1.0",
     "eslint-plugin-promise": "^2.0.1",
-    "eslint-plugin-standard": "^1.3.2",
     "grunt": "^0.4.5",
-    "grunt-bump": "^0.6.0",
     "grunt-babel": "^6.0.0",
+    "grunt-bump": "^0.6.0",
     "grunt-cli": "^1.1.0",
-    "grunt-npm": "0.0.2",
     "grunt-contrib-clean": "^1.0.0",
     "grunt-contrib-watch": "^1.0.0",
     "grunt-eslint": "^18.0.0",
-    "load-grunt-tasks": "^3.4.1",
+    "grunt-npm": "0.0.2",
     "ios-uicatalog": "^1.0.4",
+    "load-grunt-tasks": "^3.4.1",
     "mocha": "^3.0.2",
     "wdio-mocha-framework": "^0.4.0",
     "webdriverio": "^4.2.8"
